+++ airline-project/src/main/java/app/controllers/api/rest/PassengerRestApi.java
@@ -12,14 +12,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 
@@ -86,7 +79,7 @@ public interface PassengerRestApi {
             @ApiResponse(code = 200, message = "Passenger updated"),
             @ApiResponse(code = 404, message = "Passenger not found")
     })
-    @PutMapping(value = "/{id}")
+    @PatchMapping(value = "/{id}")
     ResponseEntity<PassengerDTO> update(
             @ApiParam(
                     name = "id",
@@ -98,7 +91,8 @@ public interface PassengerRestApi {
                     name = "Passenger",
                     value = "Passenger model"
             )
-            @RequestBody @Valid PassengerDTO passengerDTO);
+            @RequestBody
+            @Valid PassengerDTO passengerDTO);
 
     @ApiOperation(value = "Delete Passenger by \"id\"")
     @ApiResponses(value = {